+++ models/contacts.json
@@ -76,5 +76,11 @@
     "email": "jec@Nulla.com",
     "phone": "(099) 558-8888",
     "id": "7a86ba1d-f830-4de9-ac6c-87f7ffa28c67"
+  },
+  {
+    "name": "Juseppe Mandini",
+    "email": "kec@Nulla.com",
+    "phone": "(099) 558-7777",
+    "id": "c29bbe59-a365-4f5c-a33a-911a6c4a64ec"
   }
 ]
