+++ routes/api/contacts.js
@@ -1,98 +1,27 @@
 const express = require("express");
 
-const Joi = require("joi");
+// const Joi = require("joi");
 
 const router = express.Router();
+const ctrl = require('../../controllers/contacts')
+// const contacts = require("../../models/contacts");
 
-const contacts = require("../../models/contacts");
+// const { HttpError } = require("../../helpers");
 
-const { HttpError } = require("../../helpers");
+// const addSchema = Joi.object({
+//   name: Joi.string().required(),
+//   email: Joi.string().email().required(),
+//   phone: Joi.string().required(),
+// });
 
-const addSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().email().required(),
-  phone: Joi.string().required(),
-});
+router.get("/", ctrl.listContacts);
 
-router.get("/", async (req, res, next) => {
-  try {
-    // res.json({ message: 'template message' })
-    const result = await contacts.listContacts();
-    res.json(result);
-  } catch (error) {
-    // res.status(500).json({ message: "Server error" });
-    next(error);
-  }
-});
+router.get("/:contactId", ctrl.getContactById);
 
-router.get("/:contactId", async (req, res, next) => {
-  // res.json({ message: "template message" });
-  // console.log(req.params)
-  try {
-    const { contactId } = req.params;
-    const result = await contacts.getContactById(contactId);
-    if (!result) {
-      throw HttpError(404, "Not found");
-      // const error = new Error('Not found');
-      // error.status = 404;
-      // throw error;
-      // return res.status(404).json({ message: "Not found" });
-    }
-    res.json(result);
-  } catch (error) {
-    // const { status = 500, message = "Server Error" } = error;
-    // res.status(status).json({ message });
-    next(error);
-  }
-});
+router.post("/", ctrl.addContact);
 
-router.post("/", async (req, res, next) => {
-  // res.json({ message: "template message" });
-  try {
-    const { error } = addSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const result = await contacts.addContact(req.body);
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.delete("/:contactId", ctrl.removeContact);
 
-router.delete("/:contactId", async (req, res, next) => {
-  // res.json({ message: "template message" });
-  try {
-    const { contactId } = req.params;
-    const result = await contacts.removeContact(contactId);
-    if (!result) {
-      throw HttpError(404, "Not found");
-    }
-    res.json({
-      message: "Delete success",
-    });
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.put("/:contactId", async (req, res, next) => {
-  try {
-    const { error } = addSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-
-    const { contactId } = req.params;
-    const result = await contacts.updateContact(contactId, req.body);
-    if (!result) {
-      throw HttpError(404, "Not found");
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-  // res.json({ message: "template message" });
-});
+router.put("/:contactId", ctrl.updateContact);
 
 module.exports = router;
